+++ OctopusParent/OrderServiceProvider/src/main/java/com/octopus/eureka/order/OrderFinancialController.java
@@ -25,7 +25,7 @@ public class OrderFinancialController {
     OrderFinancialMapper orderFinancialMapper;
 
     @RequestMapping("/order/{orderSeq}")
-    public OrderFinancialDto getOrder(@PathVariable("orderSeq") float orderSeq) {
+    public OrderFinancialDto getOrder(@PathVariable("orderSeq") double orderSeq) {
         logger.info("请求参数orderSeq："+ orderSeq);
         return orderFinancialMapper.selectById(orderSeq);
     }
@@ -46,7 +46,7 @@ public class OrderFinancialController {
     }
 
     @RequestMapping("/order/delete/{orderSeq}")
-    public int deleteOrder(@PathVariable("orderSeq") float orderSeq) {
+    public int deleteOrder(@PathVariable("orderSeq") double orderSeq) {
         logger.info("请求参数orderSeq："+ orderSeq);
         return orderFinancialMapper.delete(orderSeq);
     }
